events.lua: Fix bug with unexpected Classes on TBC.
On TBC GetNumClasses() returns 11, but the expansion only has 9
classes. The code goes on to call GetClassInfo() which returns
nil for classIndex when it does not exist in the game.
i.e 6 - Death Knight, 10 - Monk, etc. It then tries
to index RAID_CLASS_COLORS with nil which of course causes an error.